linting

+++ vision/keypoints/briefdesc.go
@@ -74,9 +74,9 @@ func ComputeBRIEFDescriptors(img *image.Gray, kps *FASTKeypoints, cfg *BRIEFConf
 	}
 	// sample positions
 	xs0 := sampleIntegers(cfg.PatchSize, cfg.N, cfg.Sampling)
-	ys0 := utils.CycleIntSliceByN(sampleIntegers(cfg.PatchSize, cfg.N, cfg.Sampling), int(cfg.N/4))
-	xs1 := utils.CycleIntSliceByN(sampleIntegers(cfg.PatchSize, cfg.N, cfg.Sampling), int(cfg.N/2))
-	ys1 := utils.CycleIntSliceByN(sampleIntegers(cfg.PatchSize, cfg.N, cfg.Sampling), int(3*cfg.N/4))
+	ys0 := utils.CycleIntSliceByN(sampleIntegers(cfg.PatchSize, cfg.N, cfg.Sampling), cfg.N/4)
+	xs1 := utils.CycleIntSliceByN(sampleIntegers(cfg.PatchSize, cfg.N, cfg.Sampling), cfg.N/2)
+	ys1 := utils.CycleIntSliceByN(sampleIntegers(cfg.PatchSize, cfg.N, cfg.Sampling), 3*cfg.N/4)
 
 	// compute descriptors
 	descriptors := make(Descriptors, len(kps.Points))
